+++ lib/src/enums.dart
@@ -1,0 +1,3 @@
+library enums;
+
+part './enums/animations.dart';