+++ lib/src/enums/animations.dart
@@ -1,0 +1,12 @@
+part of enums;
+
+enum KyTransition {
+  /// [SlideTransition]
+  rightToLeft,
+  leftToRight,
+  topToBottom,
+  bottomToTop,
+
+  /// [FadeTransition]
+  fade
+}