+++ lib/src/models.dart
@@ -1,5 +1,5 @@
 library models;
 
-import 'package:koyo/src/widgets.dart';
+import 'package:koyo/src/enums.dart';
 
 part 'models/ky_page_extra.dart';