+++ lib/src/models/ky_page_extra.dart
@@ -1,16 +1,16 @@
 part of models;
 
-class KyPageExtra {
+class KyPageExtra<T extends Object> {
   KyPageExtra({
     this.transition,
     this.data,
   });
 
   factory KyPageExtra.fromExtra(Object? extra) {
-    if (extra == null || extra is! KyPageExtra) return KyPageExtra();
-    return extra;
+    if (extra != null && extra is KyPageExtra) return extra as KyPageExtra<T>;
+    return KyPageExtra();
   }
 
   KyTransition? transition;
-  Object? data;
+  T? data;
 }