+++ lib/src/widgets.dart
@@ -7,6 +7,7 @@ import 'dart:ui';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:koyo/koyo.dart';
+import 'package:koyo/src/enums.dart';
 
 part 'widgets/koyo.dart';
 part 'widgets/animations/ky_page_transition.dart';